Implemented classes for Wards and Tram stops
I've set it up so that each node adds itself to a list (I.e tram_Stops)
this can then be access  and iterated through using a foreach  to change
an aspect of all the nodes. At the moment if you press space it switches
the census data to "travelByBus" and if you press N it changes to
travelByCar (All can me mapped to buttons etc.
Also added a loading text at the bottom so you know the data is loading.